pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "scraper"
version = "3.0.0"
dependencies = [
    "pdfplumber==0.11.8",
    "requests==2.32.5",
    "pydantic==2.12.5",
    "pytest>=9.0.2",
    "typer>=0.21.1",
    "fastapi[standard]>=0.128.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
authors = [{ name = "Ming Li Liu", email = "vincentmingli@gmail.com" }]
maintainers = [{ name = "Ming Li Liu", email = "vincentmingli@gmail.com" }]
license = "MIT"
requires-python = ">=3.14"

[project.optional-dependencies]
dev = ["ruff", "mypy"]

[project.scripts]
app = "api.app:app"
scraper = "scraper.main:main"
